refactor(table): 메뉴 모듈 분리
