+++ menu/src/main/java/ktichenpos/menu/domain/MenuGroupRepository.java
@@ -1,0 +1,13 @@
+package ktichenpos.menu.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import kitchenpos.common.exception.NotFoundException;
+
+public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
+
+	default MenuGroup menuGroup(long id) {
+		return findById(id).orElseThrow(
+			() -> new NotFoundException(String.format("메뉴 그룹 id(%d)를 찾을 수 없습니다.", id)));
+	}
+}
