+++ menu/src/main/java/ktichenpos/menu/domain/MenuProductRepository.java
@@ -1,0 +1,6 @@
+package ktichenpos.menu.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MenuProductRepository extends JpaRepository<MenuProduct, Long> {
+}
