+++ menu/src/main/java/ktichenpos/menu/ui/response/MenuResponse.java
@@ -1,0 +1,64 @@
+package ktichenpos.menu.ui.response;
+
+import java.math.BigDecimal;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import kitchenpos.menu.domain.Menu;
+
+public class MenuResponse {
+
+	private long id;
+	private String name;
+	private BigDecimal price;
+	private Long menuGroupId;
+	private List<MenuProductResponse> menuProducts;
+
+	private MenuResponse() {
+	}
+
+	private MenuResponse(Long id, String name, BigDecimal price, Long menuGroupId,
+		List<MenuProductResponse> menuProducts) {
+		this.id = id;
+		this.name = name;
+		this.price = price;
+		this.menuGroupId = menuGroupId;
+		this.menuProducts = menuProducts;
+	}
+
+	public static MenuResponse from(Menu menu) {
+		return new MenuResponse(
+			menu.id(),
+			menu.name(),
+			menu.price(),
+			menu.menuGroup().id(),
+			MenuProductResponse.listFrom(menu.menuProducts())
+		);
+	}
+
+	public static List<MenuResponse> listFrom(List<Menu> menus) {
+		return menus.stream()
+			.map(MenuResponse::from)
+			.collect(Collectors.toList());
+	}
+
+	public Long getId() {
+		return id;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	public BigDecimal getPrice() {
+		return price;
+	}
+
+	public Long getMenuGroupId() {
+		return menuGroupId;
+	}
+
+	public List<MenuProductResponse> getMenuProducts() {
+		return menuProducts;
+	}
+}
