+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name = 'kitchenpos'
-include 'table', 'common'
+include 'table', 'common', 'menu'
