--- src/main/java/kitchenpos/menu/ui/MenuRestController.java
@@ -1,38 +1,0 @@
-package kitchenpos.menu.ui;
-
-import java.net.URI;
-import java.util.List;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-import kitchenpos.menu.application.MenuService;
-import kitchenpos.menu.ui.request.MenuRequest;
-import kitchenpos.menu.ui.response.MenuResponse;
-
-@RestController
-public class MenuRestController {
-    private final MenuService menuService;
-
-    public MenuRestController(final MenuService menuService) {
-        this.menuService = menuService;
-    }
-
-    @PostMapping("/api/menus")
-    public ResponseEntity<MenuResponse> create(@RequestBody final MenuRequest request) {
-        final MenuResponse created = menuService.create(request);
-        final URI uri = URI.create("/api/menus/" + created.getId());
-        return ResponseEntity.created(uri)
-                .body(created)
-                ;
-    }
-
-    @GetMapping("/api/menus")
-    public ResponseEntity<List<MenuResponse>> list() {
-        return ResponseEntity.ok()
-                .body(menuService.list());
-    }
-}
