--- src/main/java/kitchenpos/menu/ui/ProductRestController.java
@@ -1,36 +1,0 @@
-package kitchenpos.menu.ui;
-
-import java.net.URI;
-import java.util.List;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-import kitchenpos.menu.application.ProductService;
-import kitchenpos.menu.ui.request.ProductRequest;
-import kitchenpos.menu.ui.response.ProductResponse;
-
-@RestController
-public class ProductRestController {
-    private final ProductService productService;
-
-    public ProductRestController(final ProductService productService) {
-        this.productService = productService;
-    }
-
-    @PostMapping("/api/products")
-    public ResponseEntity<ProductResponse> create(@RequestBody final ProductRequest request) {
-        final ProductResponse created = productService.create(request);
-        final URI uri = URI.create("/api/products/" + created.getId());
-        return ResponseEntity.created(uri).body(created);
-    }
-
-    @GetMapping("/api/products")
-    public ResponseEntity<List<ProductResponse>> list() {
-        return ResponseEntity.ok()
-            .body(productService.list());
-    }
-}
